[refactor] - 각 패키지 정리
