+++ src/main/java/com/yanus/attendance/attendance/application/workschedule/WorkScheduleEventService.java
@@ -1,10 +1,9 @@
-package com.yanus.attendance.attendance.application;
+package com.yanus.attendance.attendance.application.workschedule;
 
 import com.yanus.attendance.attendance.domain.workschedule.WorkScheduleEvent;
 import com.yanus.attendance.attendance.domain.workschedule.WorkScheduleEventRepository;
 import com.yanus.attendance.attendance.presentation.dto.WorkScheduleEventRequest;
 import com.yanus.attendance.attendance.presentation.dto.WorkScheduleEventResponse;
-import com.yanus.attendance.attendance.presentation.dto.WorkScheduleResponse;
 import com.yanus.attendance.global.exception.BusinessException;
 import com.yanus.attendance.global.exception.ErrorCode;
 import com.yanus.attendance.member.domain.Member;
